removed MongoDB.py file and moved the code to app.py
